feat: Introduce artificial delay in App A readiness probe
To demonstrate the dependency problem, an initialDelaySeconds of 30 has been
added to the readinessProbe of App A (Redis). This will cause App A to report
itself as ready to Kubernetes only after 30 seconds, even if the Redis process
starts sooner. This should allow App B to start and attempt connections to
Redis before it is truly ready, thus demonstrating the problem.